Move extra scopes and extra qp to interactive parameters
