+++ common/src/main/java/com/microsoft/identity/common/internal/commands/parameters/TokenCommandParameters.java
@@ -22,8 +22,6 @@
 // THE SOFTWARE.
 package com.microsoft.identity.common.internal.commands.parameters;
 
-import android.util.Pair;
-
 import com.google.gson.annotations.Expose;
 import com.microsoft.identity.common.exception.ArgumentException;
 import com.microsoft.identity.common.internal.authorities.Authority;
@@ -31,10 +29,8 @@ import com.microsoft.identity.common.internal.authscheme.AbstractAuthenticationS
 import com.microsoft.identity.common.internal.dto.IAccountRecord;
 import com.microsoft.identity.common.internal.logging.Logger;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
 
 import lombok.EqualsAndHashCode;
@@ -47,10 +43,6 @@ import lombok.experimental.SuperBuilder;
 public class TokenCommandParameters extends CommandParameters {
 
     private static final String TAG = TokenCommandParameters.class.getSimpleName();
-
-    private List<Pair<String, String>> extraQueryStringParameters;
-
-    private List<String> extraScopesToConsent;
 
     private IAccountRecord account;
 
@@ -66,18 +58,9 @@ public class TokenCommandParameters extends CommandParameters {
     @Expose()
     private AbstractAuthenticationScheme authenticationScheme;
 
-    public List<Pair<String, String>> getExtraQueryStringParameters() {
-        return this.extraQueryStringParameters == null ? null : new ArrayList<>(this.extraQueryStringParameters);
-    }
-
-    public List<String> getExtraScopesToConsent() {
-        return this.extraScopesToConsent == null ? null : new ArrayList<>(this.extraScopesToConsent);
-    }
-
     public Set<String> getScopes() {
         return this.scopes == null ? null : new HashSet<>(this.scopes);
     }
-
 
     public void validate() throws ArgumentException {
         final String methodName = ":validate";
